+++ app/api/v1/data/trajectory/route.ts
