+++ app/data/compare/page.tsx
@@ -2,13 +2,17 @@
 
 import DataCompare from "@/components/DataCompare";
 import { Suspense } from "react";
+import { useSessionCheck } from "@/hooks/useSessionCheck";
 
-const Page = () => {
+export default function Page() {
+  const { sessionCheck } = useSessionCheck();
+
+  if (sessionCheck) {
+    return <div>Loading...</div>;
+  }
   return (
     <Suspense>
       <DataCompare />
     </Suspense>
   );
-};
-
-export default Page;
+}
