+++ hooks/useCameraData.ts
